Add GPU (CUDA) support for KSC.
When `leuven` is built with CUDA support, i.e. by using the -DUSE_CUBLAS=ON
CMake configuration option, GPU can be used in the Training part of the KSC.
This includes the QR, SVD decompostions and computation of QU, i.e. the
required eiegenvectors of the coresponding symmetric problem. These computations
can be accelerated by exploiting GPU (cuBLAS, cuSOLVER and some CUDA kernels as
glue) especially in case of large size of reduced sets, when the SVD can take
significant time on the CPU. (Note, that this SVD is multi threaded only when
using intel MKL, i.e. no way of accelerate it on MAD machines other than
performing it on GPU.)